+++ odoo/local-src/sf_account_analytic/__manifest__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+# Part of senseFly.
+
+{
+    'name': 'SenseFly Account Analytic',
+    'version': '1.0',
+    'author': 'Telmo Santos',
+    'category': 'SenseFly',
+    'depends': ['analytic'],
+    'data': ['views/account_analytic.xml'],
+    'demo': [],
+    'images': ['static/description/icon.png'],
+    'installable': True,
+    'auto_install': False,
+}