+++ HISTORY.rst
@@ -22,6 +22,7 @@ latest (unreleased)
 **Features and Improvements**
 
 * Add active field on journal
+* Add field owner in account analytic
 * Add Partner: Customer, supplier (draft)
 * Add the import Sales Team / Channel to demo and install songs
 * Add the import "Drone type" to demo and install songs